[JSSC-67] 🔨 Refactor : UploadPostHeader 컴포넌트 분리
